G3: Split TodayFeed offline/fallback composite widget into sub-widgets
- Deleted legacy states/offline_state_widget.dart (789 LOC)
- Created new folder features/today_feed/presentation/widgets/offline/
  • offline_state_widget.dart (≈320 LOC)
  • fallback_state_widget.dart (≈340 LOC)
  • components/status_badge.dart, topic_badge.dart
- Updated TodayFeedTile imports
- Updated docs tech plan: mark G3 ✅

All tests & analyzer green.
