+++ app/lib/features/today_feed/presentation/widgets/offline/components/status_badge.dart
@@ -1,0 +1,48 @@
+import 'package:flutter/material.dart';
+import '../../../../../../core/services/responsive_service.dart';
+
+/// A small pill-shaped badge showing an icon + label.
+class StatusBadge extends StatelessWidget {
+  const StatusBadge({
+    super.key,
+    required this.text,
+    required this.icon,
+    required this.color,
+  });
+
+  final String text;
+  final IconData icon;
+  final Color color;
+
+  @override
+  Widget build(BuildContext context) {
+    final spacing = ResponsiveService.getTinySpacing(context);
+    final iconSize = ResponsiveService.getIconSize(context, baseSize: 12);
+
+    return Container(
+      padding: EdgeInsets.symmetric(horizontal: spacing * 2, vertical: spacing),
+      decoration: BoxDecoration(
+        color: color.withValues(alpha: 0.1),
+        borderRadius: BorderRadius.circular(
+          ResponsiveService.getBorderRadius(context) / 2,
+        ),
+        border: Border.all(color: color.withValues(alpha: 0.3), width: 1),
+      ),
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Icon(icon, size: iconSize, color: color),
+          SizedBox(width: spacing),
+          Text(
+            text,
+            style: Theme.of(context).textTheme.labelSmall?.copyWith(
+              color: color,
+              fontWeight: FontWeight.w600,
+              letterSpacing: 0.5,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
